move containers into containers folder and rename paths accordingly
