+++ client/components/Bill/Bill.js
@@ -4,7 +4,7 @@ import { Link } from 'react-router-dom'
 import CommentCount from '../CommentCount/CommentCount'
 
 import './bill.css'
-import Vote from '../Vote/Vote.js'
+import Vote from '../../containers/Vote/Vote'
 
 const Bill = (props) => {
   let title = props.title
