+++ client/components/Home/Home.js
@@ -1,8 +1,8 @@
 import React from 'react'
 
 import Header from '../Header/Header'
-import About from '../About/About'
-import Bills from '../Bills/Bills'
+import About from '../../containers/About/About'
+import Bills from '../../containers/Bills/Bills'
 import Footer from '../Footer/Footer'
 
 import './home.css'
