+++ client/containers/App.jsx
@@ -2,7 +2,7 @@ import React, { Component } from 'react'
 import { Route } from 'react-router'
 
 import Home from '../components/Home/Home'
-import BillPage from '../components/BillPage/BillPage'
+import BillPage from './BillPage/BillPage'
 
 class App extends Component {
   render () {
