+++ client/containers/Vote/Vote.js
