+++ client/containers/Vote/vote.css
